move utils/ManagedTree to shared/ManagedTree
