+++ src/components/SourcesTree.js
@@ -11,7 +11,7 @@ const {
   nodeHasChildren, createParentMap, addToTree,
   collapseTree, createTree, getDirectories
 } = require("../utils/sources-tree.js");
-const ManagedTree = React.createFactory(require("./utils/ManagedTree"));
+const ManagedTree = React.createFactory(require("./shared/ManagedTree"));
 const actions = require("../actions");
 const Svg = require("./shared/Svg");
 const { throttle } = require("../utils/utils");
