+++ src/components/shared/ObjectInspector.js
@@ -1,6 +1,6 @@
 const React = require("react");
 const classnames = require("classnames");
-const ManagedTree = React.createFactory(require("../utils/ManagedTree"));
+const ManagedTree = React.createFactory(require("./ManagedTree"));
 const Svg = require("./Svg");
 const Rep = require("./Rep");
 const { MODE } = require("devtools-reps");
